% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAT.R
\name{calcAT}
\alias{calcAT}
\title{Calculate Apparent Temperature (AT)}
\usage{
calcAT(tdb, rh, v, q = NULL, round = TRUE)
}
\arguments{
\item{tdb}{Dry bulb air temperature in degrees Celsius.}

\item{rh}{Relative humidity in percent.}

\item{v}{Wind speed 10m above ground level in meters per second.}

\item{q}{Net radiation absorbed per unit area of body surface in W/m2.}

\item{round}{If TRUE, the output value is rounded to one decimal place.}
}
\value{
Apparent temperature in degrees Celsius.
}
\description{
This function calculates the Apparent Temperature (AT), which
is defined as the temperature at the reference humidity level
producing the same amount of discomfort as that experienced under
the current ambient conditions of temperature, humidity,
and possibly solar radiation.
}
\examples{
at(tdb = 25, rh = 30, v = 0.1) # Without solar radiation
at(tdb = 25, rh = 50, v = 0.1, q = 300) # With solar radiation
}
