% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcATHBpmv.r
\name{calcATHBpmv}
\alias{calcATHBpmv}
\title{Calculation of Adaptive Thermal Heat Balance Indices with PMV}
\description{
\code{calcATHBpmv} calculates three different indices related to the adaptive thermal heat balance framework according to PMV
}
\usage{
calcATHBpmv(trm, psych, ta, tr, vel, rh, met, wme)
}
\arguments{
\item{trm}{- Running mean outdoor temperature in [degree C]}

\item{psych}{- factor related to fixed effect on perceived control}

\item{ta}{- a numeric value presenting air temperature in [degree C]}

\item{tr}{- a numeric value presenting mean radiant temperature in [degree C]}

\item{vel}{- a numeric value presenting air velocity in [m/s]}

\item{rh}{- a numeric value presenting relative humidity [\%]}

\item{met}{- a numeric value presenting metabolic rate in [met]}

\item{wme}{- a numeric value presenting external work in [met]}
}
\details{
All variables must have the same length 1. For the calculation of several values use function \code{calcComfInd}.
}
\value{
pmv value adapted through the ATHB approach
}
\description{
calcATHB calculates three different indices related to the adaptive thermal heat balance framework. This function returns the pmv value adapted through the ATHB approach
}
\references{
Schweiker, M. & Wagner, A. A framework for an adaptive thermal heat balance model (ATHB), Building and Environment, 2015, 94, 252 - 262

Schweiker, M. & Wagner, A. Exploring potentials and limitations of the adaptive thermal heat balance framework Proceedings of 9th Windsor Conference: making comfort relevant Cumberland Lodge, Windsor, UK, 2016
}
\author{
Marcel Schweiker
}

\seealso{
see also \code{\link{calcComfInd}}
}
\examples{
calcATHBpmv(20, 0, 25, 25, .1, 50, 1.1, 0)

}
