% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAVAbloodflow.R
\name{calcAVAbloodflow}
\alias{calcAVAbloodflow}
\alias{calcEvap}
\title{Calculate areteriovenous anastmoses (AVA) blood flow rate [L/h]}
\usage{
calcEvaporation(err_cr, err_sk, tsk, ta, rh, ret, height=1.72, weight=74.43, equation="dubois", age=20)

calcEvap(err_cr, err_sk, tsk, ta, rh, ret, height=1.72, weight=74.43, equation="dubois", age=20)
}
\arguments{
\item{err_cr}{Float or array, optional. Difference between setpoint and body temperatures.The default is 0.}

\item{err_sk}{Float or array, optional. Difference between setpoint and body temperatures.The default is 0.}

\item{height}{Body height [m]. The default is 1.72.}

\item{weight}{Body weight [kg]. The default is 74.43.}

\item{age}{Age [years]. The default is 20.}

\item{tsk}{Skin temperatures in [Degree Celcius]}

\item{ta}{Air temperatures at local body segments in [Degree Celcius]}

\item{rh}{Relative humidity at local body segments [%]}

\item{ret}{Total evaporative thermal resistances [m2.K/W]}
}
\value{
BFava_hand: AVA blood flow rate at hand [L/h].

BFava_foot : AVA blood flow rate at foot [L/h].
}
\description{
Function to Calculate areteriovenous anastmoses (AVA) blood flow rate [L/h] based on Takemori's model, 1995.
}
\seealso{
\code{\link{helpJOS}}
}
\author{
Code implemented in to R by Shoaib Sarwar. Further contribution by Marcel Schweiker.
}
