% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBFBrate.R
\name{calcBFBrate}
\alias{calcBFBrate}
\alias{calcbfbrate}
\alias{bfbrate}
\title{Calculate the ratio of basal blood flow}
\usage{
calcBFBrate(height=1.72, weight=74.43, age=20, ci=2.59, equation="dubois")

calcbfbrate(height=1.72, weight=74.43, age=20, ci=2.59, equation="dubois")

bfbrate(height=1.72, weight=74.43, age=20, ci=2.59, equation="dubois")
}
\arguments{
\item{height}{Body height. The default is 1.72 in [m]}

\item{weight}{Body height. The default is 74.43 in [kg]}

\item{age}{Age. The default is 20 in [years]}

\item{ci}{Cardiac index. The default is 2.59 in [L/min/m2]}

\item{equation}{The equation name of bsa calculation. The default is "dubois" in [str]}
}
\value{
Basal blood flow rate (bfb_rate)
}
\description{
Function to Calculate the ratio of basal blood flow (BFB) of the standard body (290 L/h).
}
\details{
Choose a name from "dubois","takahira", "fujimoto", or "kurazumi". All these equations are defined in helpJOS functions.
}
\examples{
calcBFBrate()
}
\seealso{
\code{\link{helpJOS}}
}
\author{
Code implemented in to R by Shoaib Sarwar. Further contribution by Marcel Schweiker.
}
