% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBSA.R
\name{calcBSA}
\alias{calcBSA}
\alias{calcbsa}
\alias{bsa}
\title{Calculate Body Surface Area}
\usage{
calcBSA(height=1.72, weight=74.43, equation="dubois")

calcbsa(height=1.72, weight=74.43, equation="dubois")

bsa(height=1.72, weight=74.43, equation="dubois")
}
\arguments{
\item{height}{Body height. The default is 1.72 in [m]}

\item{weight}{Body height. The default is 74.43 [kg]}

\item{equation}{The equation name of bsa calculation. The default is "dubois" in [str]}
}
\value{
Body surface area (BSA) [m2]
}
\description{
Function to calculate body surface area (BSA) in [m2]
}
\details{
Choose a name from "dubois","takahira", "fujimoto", or "kurazumi". All these equations are defined in helpJOS functions.
}
\examples{
calcBSA()
}
\seealso{
\code{\link{helpJOS}}
}
\author{
Code implemented in to R by Shoaib Sarwar. Further contribution by Marcel Schweiker.
}
