% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBSAlocal.R
\name{calcBSAlocal}
\alias{calcBSAlocal}
\alias{calcbsalocal}
\alias{bsalocal}
\title{Calculate Local Body Surface Area}
\usage{
calcBSAlocal(height=1.72, weight=74.43, equation="dubois")

calcbsalocal(height=1.72, weight=74.43, equation="dubois")

bsalocal(height=1.72, weight=74.43, equation="dubois")
}
\arguments{
\item{height}{Body height. The default is 1.72 in [m]}

\item{weight}{Body height. The default is 74.43 [kg]}

\item{equation}{The equation name of bsa calculation. The default is "dubois" in [str]}
}
\value{
Local body surface area (BSA) [m2].
}
\description{
Function to  Calculate local body surface area (BSA) [m2].
}
\details{
Choose a name from "dubois","takahira", "fujimoto", or "kurazumi". All these equations are defined in helpJOS functions.
}
\examples{
calcBSAlocal()
}
\seealso{
\code{\link{helpJOS}}
}
\author{
Code implemented in to R by Shoaib Sarwar. Further contribution by Marcel Schweiker.
}
