% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBSArate.R
\name{calcBSArate}
\alias{calcBSArate}
\alias{calcbsarate}
\alias{bsarate}
\title{Calculate the rate of BSA to standard body}
\usage{
calcBSArate(height=1.72, weight=74.43, equation="dubois")

calcbsarate(height=1.72, weight=74.43, equation="dubois")

bsarate(height=1.72, weight=74.43, equation="dubois")
}
\arguments{
\item{height}{Body height. The default is 1.72 in [m]}

\item{weight}{Body height. The default is 74.43 [kg]}

\item{equation}{The equation name of bsa calculation. The default is "dubois" in [str]}
}
\value{
ratio of BSA to the standard body [-]
}
\description{
Function to calculate the ratio of BSA to the standard body.
}
\details{
Choose a name from "dubois","takahira", "fujimoto", or "kurazumi". All these equations are defined in helpJOS functions.
}
\examples{
calcBSArate()
}
\seealso{
\code{\link{helpJOS}}
}
\author{
Code implemented in to R by Shoaib Sarwar. Further contribution by Marcel Schweiker.
}
