% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCloTout.R
\name{calcCloTout}
\alias{calcCloTout}
\title{Representative clothing insulation Icl as a function of outdoor air temperature at 06:00 a.m [4].}
\usage{
calcCloTout(tout, units = "SI")
}
\arguments{
\item{tout}{float, int, or array-like outdoor air temperature at 06:00 a.m.
default in [Celsius] in [Fahrenheit] if 'units' = 'IP'}

\item{units}{is \{"SI", "IP"\}, select the SI \(International System of Units\) or 
the IP (Imperial Units) system. Defaults to 'SI'.}
}
\value{
clo : float, int, or array-like. Representative clothing insulation Icl, [clo]
}
\description{
Representative clothing insulation Icl as a function of outdoor air temperature at 06:00 a.m [4].
}
\details{
The ASHRAE 55 2020 states that it is acceptable to determine the clothing insulation 
Icl using this equation in mechanically conditioned buildings [1]_. Limitations: This equation 
may not be accurate for extreme temperature ranges.
}
\examples{
calcCloTout(tout=27)
}
\references{
ASHRAE 55 2020
}
\author{
Code implemented into R by Yiqing Zhang.
}
