% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDiscomfortIndex.R
\name{calcDiscomfortIndex}
\alias{calcDiscomfortIndex}
\title{Calculates the Discomfort Index(DI).
The Discomfort Index is based on air temperature and humidity.
Thermal discomfort prevention strategy.It is compliant with ISO 15265.}
\usage{
calcDiscomfortIndex(tdb, rh)
}
\arguments{
\item{tdb}{a numeric value presenting air temperature in [degree C].}

\item{rh}{a numeric value presenting relative humidity.}
}
\value{
A list containing:
 \itemize{
   \item \code{di}: Numeric value(s) of the Discomfort Index (DI)
    in degrees Celsius, rounded to 1 decimal place.
   \item \code{discomfort_condition}: Character vector or list representing
   the discomfort category based on the DI value(s), such as "No discomfort"
   or "State of medical emergency".
 }
}
\description{
Calculates the Discomfort Index(DI).
The Discomfort Index is based on air temperature and humidity.
Thermal discomfort prevention strategy.It is compliant with ISO 15265.
}
\details{
The Discomfort Index (DI) is an effective temperature based
on air temperature and humidity, typically used for warm environments.
It is divided into six categories:
[1]: DI < 21 Celsius - No discomfort
[2]: 21 Celsius <= DI < 24 Celsius - Less than 50 feel discomfort
[3]: 24 Celsius <= DI < 27 Celsius - More than 50 feel discomfort
[4]: 27 Celsius <= DI < 29 Celsius - Most people feel discomfort
[5]: 29 Celsius <= DI < 32 Celsius - Severe stress for everyone
[6]: DI >= 32 Celsius - Medical emergency
}
\examples{
calcDiscomfortIndex(tdb=25, rh=50)
}
\references{
ISO 15265
}
\author{
Code implemented into R by Haiyang Liu.
}
