% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHeatIndex.R
\name{calcHeatIndex}
\alias{calcHeatIndex}
\title{Calculates the Heat Index (HI).}
\usage{
calcHeatIndex(tdb, rh, units = "SI")
}
\arguments{
\item{tdb}{float value that dry bulb air temperature, default in Celsius in Fahrenheit if 'units' = 'IP'}

\item{rh}{float value that is relative humidity}

\item{units}{is {'SI', 'IP'}, select the SI (International System of Units) or the IP (Imperial Units) system.}
}
\value{
hi, a float value, is the Heat Index, default in Celsius inFahrenheit if 'units' = 'IP'
}
\description{
Calculates the Heat Index (HI).
}
\details{
It combines air temperature and relative humidity to determine an apparent temperature. 
The HI equation [12]_ is derived by multiple regression analysis in temperature and relative 
humidity from the first version of Steadman's (1979) apparent temperature (AT) [13]_.
}
\examples{
calcHeatIndex(tdb=25, rh=50)
}
\references{
Steadman's (1979) apparent temperature (AT)
}
\author{
Code implemented into R by Yiqing Zhang.
}
