% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPetSteady.R
\name{calcPetSteady}
\alias{calcPetSteady}
\title{Calculate the Steady Physiological Equivalent Temperature (PET)}
\usage{
calcPetSteady(
  tdb,
  tr,
  v,
  rh,
  met,
  clo,
  p_atm = 1013.25,
  position = 1,
  age = 23,
  sex = 1,
  weight = 75,
  height = 1.8,
  wme = 0
)
}
\arguments{
\item{tdb}{float, Dry bulb air temperature in degrees Celsius.}

\item{tr}{float, Mean radiant temperature in degrees Celsius.}

\item{v}{float, Air speed in meters per second (m/s).}

\item{rh}{float, Relative humidity in percent (\%).}

\item{met}{float, Metabolic rate in metabolic rate units (met).}

\item{clo}{float, Clothing insulation in clo units.}

\item{p_atm}{float, Atmospheric pressure in hectopascals (hPa), default value 1013.25.}

\item{position}{integer, Position of the individual: 1 = sitting, 2 = standing,
3 = standing with forced convection.}

\item{age}{integer, Age of the individual in years, default 23.}

\item{sex}{integer, Gender of the individual: 1 for male, 2 for female, default 1.}

\item{weight}{float, Body mass in kilograms (kg), default 75.}

\item{height}{float, Height of the individual in meters (m), default 1.8.}

\item{wme}{float, External work performed by the individual in watts per square meter
(W/m^2), default 0.}
}
\value{
Returns the calculated steady-state Physiological Equivalent Temperature (PET) under
the specified ambient conditions, in degrees Celsius.
}
\description{
Calculates the steady-state Physiological Equivalent Temperature (PET) using the
Munich Energy-balance Model for Individuals. PET is defined as the air temperature at
which, in a typical indoor setting, the heat budget of the human body is balanced,
maintaining the same core and skin temperature that would be achieved under the assessed
outdoor conditions.

The PET index was originally proposed by Hoppe and later revised by Walther and Goestchel
in 2018, who corrected errors in the initial PET calculation routine and implemented a
state-of-the-art vapor diffusion model.
}
\examples{
# Calculate PET for typical indoor conditions
calcPetSteady(tdb = 20, tr = 20, rh = 50, v = 0.15, met = 1.37, clo = 0.5)

}
\references{
Hoppe, P. (1999). The Physiological Equivalent Temperature - A Universal Index for the
Biometeorological Assessment of the Thermal Environment.
Walther, G., and Goestchel, P. (2018). Improvement and Extension of the Munich
Energy-balance Model for Individuals.
}
\author{
Code implemented into R by Chongyu Gan.
}
