% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWbgt.R
\name{calcWbgt}
\alias{calcWbgt}
\title{Calculates the Wet Bulb Globe Temperature (WBGT)}
\usage{
calcWbgt(twb, tg, tdb = NULL, with_solar_load = FALSE, round = TRUE)
}
\arguments{
\item{twb}{Numeric value representing natural (no forced air flow)
wet bulb temperature in degrees Celsius.}

\item{tg}{Numeric value representing globe temperature in [degree C].}

\item{tdb}{Numeric value representing dry bulb air temperature in [degree C].}

\item{with_solar_load}{Logical, TRUE if the globe sensor is exposed to sun.}

\item{round}{Logical, TRUE if output should be rounded. Default is TRUE.}
}
\value{
Numeric value of the Wet Bulb Globe Temperature Index in [degree C].
}
\description{
The WBGT is a heat stress index that measures the thermal
environment to which a person is exposed. It is compliant with ISO 7243.
}
\details{
The WBGT determines the impact of heat on a person throughout
the course of a working day (up to 8 hours). It does not apply to very brief
heat exposures. It pertains to the evaluation of male and female individuals
who are fit for work in both indoor and outdoor occupational environments,
as well as other types of environments. The WBGT is defined as
a function of only twb and tg if the person is not exposed to direct radiant
heat from the sun. When exposed, tdb must also be specified.
}
\examples{
calcWbgt(twb = 25, tg = 32, tdb = 20, with_solar_load = TRUE)
}
\references{
ISO 7243
}
\author{
Code implemented into R by Chongyu Gan.
}
