% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfASHRAETableG11.R
\docType{data}
\name{dfASHRAETableG11}
\alias{dfASHRAETableG11}
\title{Callibration data for SET}
\format{
A data frame with 22 rows and 11 variables:
\describe{
 \item{\code{ta}}{a numeric vector of air temperature [degree C]}
\item{\code{taF}}{a numeric vector of air temperature [degree F]}
 \item{\code{tr}}{a numeric vector of radiant temperature [degree C]}
\item{\code{trF}}{a numeric vector of radiant temperature [degree F]}
\item{\code{vel}}{a numeric vector of indoor air velocity [m/s]}
\item{\code{velFPM}}{a numeric vector of indoor air velocity [fpm]}
 \item{\code{rh}}{a numeric vector of relative humidity [\%]}
\item{\code{met}}{a numeric vector of metabolic rate [MET]}
 \item{\code{clo}}{a numeric vector of clothing insulation level [CLO]}
 \item{\code{set}}{a numeric vector of standard effective temperature (SET) [degree C]}
 \item{\code{setF}}{a numeric vector of standard effective temperature (SET) [degree F]}   
}
}
\usage{
data(dfASHRAETableG11)
}
\description{
Data from ASHRAE 55-2013 to callibrate values given by SET model
}
\examples{
data(dfASHRAETableG11)
head(dfASHRAETableG11)
}
\references{
ASHRAE standard 55-2013. Thermal environmental conditions for human occupancy. American society of heating, Refrigerating and Air-Conditioning Engineering, Atlanta, Usa, 2013.
}
\keyword{datasets}
