% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfISO7730AppE.R
\docType{data}
\name{dfISO7730AppE}
\alias{dfISO7730AppE}
\title{Callibration data for PMV}
\format{
A data frame with 13 rows and 8 variables:
\describe{
 \item{\code{top}}{a numeric vector of operative temperature [degree C]}
\item{\code{vel}}{a numeric vector of indoor air velocity [m/s]}
 \item{\code{rh}}{a numeric vector of relative humidity [\%]}
\item{\code{met}}{a numeric vector of metabolic rate [MET]}
 \item{\code{clo}}{a numeric vector of clothing insulation level [CLO]}
 \item{\code{pmv}}{a numeric vector of Predicted mean vote (PMV)} 
}
}
\usage{
data(dfISO7730AppE)
}
\description{
Data from ISO 7730 Appendix E to callibrate values given by PMV model
}
\examples{
data(dfISO7730AppE)
head(dfISO7730AppE)
}
\references{
ISO 7730 Ergonomics of the thermal environment analytical determination and interpretation of thermal comfort using calculation of the pmv and ppd indices and local thermal comfort criteria 2005.
}
\keyword{datasets}
