% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfISO7730TableD1.R
\docType{data}
\name{dfISO7730TableD1}
\alias{dfISO7730TableD1}
\title{Callibration data for PMV and PPD}
\format{
A data frame with 13 rows and 8 variables:
\describe{
 \item{\code{ta}}{a numeric vector of air temperature [degree C]}
 \item{\code{tr}}{a numeric vector of radiant temperature [degree C]}
\item{\code{vel}}{a numeric vector of indoor air velocity [m/s]}
 \item{\code{rh}}{a numeric vector of relative humidity [\%]}
\item{\code{met}}{a numeric vector of metabolic rate [MET]}
 \item{\code{clo}}{a numeric vector of clothing insulation level [CLO]}
 \item{\code{pmv}}{a numeric vector of Predicted mean vote (PMV)}
 \item{\code{ppd}}{a numeric vector of Predicted percentage dissatisfied (PPD)} 
}
}
\usage{
data(dfISO7730TableD1)
}
\description{
Data from ISO 7730 to callibrate values given by PMV / PPD model
}
\examples{
data(dfISO7730TableD1)
head(dfISO7730TableD1)
}
\references{
ISO 7730 Ergonomics of the thermal environment analytical determination and interpretation of thermal comfort using calculation of the pmv and ppd indices and local thermal comfort criteria 2005.
}
\keyword{datasets}
