% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfVariables.R
\docType{data}
\name{dfVariables}
\alias{dfVariables}
\title{List of all the variables used}
\format{
A data frame with 13 rows and 5 variables:
\describe{
 \item{\code{Name}}{a string representing name of the variable}
\item{\code{Unit}}{a string representing Unit of the variable}
 \item{\code{Min}}{a numeric value representing the minimum value of the variable}
\item{\code{Max}}{a numeric value representing the maximum value of the variable}
\item{\code{Definition}}{a sentence defining the variable with Unit}
}
}
\usage{
data(dfVariables)
}
\description{
List of all the variables used
}
\examples{
data(dfVariables)
head(dfVariables)
}
\keyword{datasets}
\keyword{internal}
