% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jos3.R
\name{jos3}
\alias{jos3}
\alias{simulate}
\alias{Simulate}
\alias{Jos3}
\alias{JOS3}
\alias{resetSetpt}
\alias{resetsetpt}
\title{Create JOS3 model.}
\usage{
jos3(height, weight, fat, age, sex, ci, bmrEquation, bsaEquation, exOutput, modelName)
}
\arguments{
\item{height}{Body height. The default is 1.72[m]}

\item{weight}{Body weight [kg]}

\item{fat}{Fat rate [%]}

\item{age}{Age [years]}

\item{sex}{Sex ("male" or "female")}

\item{ci}{Cardiac index [L/min/m2]}

\item{bmrEquation}{Choose a BMR equation.}

\item{bsaEquation}{Choose a BSA equation.}

\item{exOutput}{Extra output parameters. If "all", all parameters are output.}
}
\value{
the object with all the updated variables after the simulation. [list]
}
\description{
Function to create JOS3 model. The returned object can be saved and used for another simulation.
}
\details{
Creates the model on which all the other functions of JOS model will be executed. Returns the result and updated variables of the model.
}
\examples{
model1 <- jos3()
model2 <- jos3(bmrEquation <- "harris-benedict_origin")
}
\seealso{
\code{\link{resetSetpt}}

\code{\link{simulate}}
}
\author{
Code implemented in to R by Shaomi Rahman. Further contribution by Marcel Schweiker.
}
