% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resetSetpt.R
\name{resetSetpt}
\alias{resetSetpt}
\alias{simulate}
\alias{Simulate}
\alias{jos3}
\alias{Jos3}
\alias{JOS3}
\alias{resetsetpt}
\title{Set operative temperature under PMV=0 environment}
\usage{
resetSetpt(object)
}
\arguments{
\item{object}{dictionary of all the necessary variables[list]}
}
\value{
dictout a dictionary \code{dictout} the result and \code{updatedVariables} the updated variables of the model after reset.
}
\description{
Function to reset operative temperature under PMV=0 environment.
}
\details{
At first, resets the value of met, to, RH, va, icl, PAR, ava_zero. Then returns dictout with contains all the basic variable of JOS setup and the list of all the updated variables of the model.
}
\examples{
model1 = jos3()$object
resetSetpt(model1)
model2 <- jos3(bmrEquation <- "harris-benedict_origin")$object
resetSetpt(model2)$object
}
\seealso{
\code{\link{resetSetpt}}

\code{\link{jos3}}
}
\author{
Code implemented in to R by Shaomi Rahman. Further contribution by Marcel Schweiker.
}
