% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\alias{Simulate}
\alias{jos3}
\alias{Jos3}
\alias{JOS3}
\alias{resetSetpt}
\alias{resetsetpt}
\title{Execute JOS3 model.}
\usage{
simulate(times, dtime, output, object)
}
\arguments{
\item{times}{Number of loops of a simulation[numerical]}

\item{dtime}{Time delta. The default is 60[sec]}

\item{output}{If you don't record paramters, set False. The default is True. It is optional.[Boolean]}

\item{object}{The instance of JOS3 model in a list form on which JOS3 will be executed.[list]}
}
\value{
the object with all the updated variables after the simulation. [list]
}
\description{
Function to simulate JOS3 model. The returned object can be saved and used for another simulation.
}
\details{
executes the run function which executes the JOS3 model number of times as times parameter. At first, the JOS3 model must be created to create a model with jos3() function.
}
\examples{
model1 <- jos3()$object
simulate(2, 60, TRUE, model)
model2 <- jos3(bmrEquation <- "harris-benedict_origin")$object
simulate(times = 2, object = model)
}
\seealso{
\code{\link{resetSetpt}}

\code{\link{jos3}}
}
\author{
Code implemented in to R by Shaomi Rahman. Further contribution by Marcel Schweiker.
}
