% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarGainValue.R
\docType{data}
\name{solarGainValues}
\alias{solarGainValues}
\title{Dataset with DIfferent Combination of Inputs to Calculate Solar Gain}
\format{
A data frame with 384 rows and 10 variables:
\describe{
 \item{\code{solAlt}}{a numeric value presenting Solar altitude, degrees from horizontal in [degree C]}
\item{\code{solAzi}}{a numeric value presenting Solar azimuth, degrees clockwise from North in [degree C]}
 \item{\code{solRadDir}}{a numeric value presenting Direct-beam solar radiation [W/m2]}
\item{\code{solTrans}}{a numeric value presenting Total solar transmittance. Ranges from 0 to 1}
\item{\code{fSvv}}{a numeric value presenting fraction of sky vault exposed to body, ranges from 0 to 1}
\item{\code{fBes}}{a numeric value presenting fraction of the possible body surface exposed to sun, ranges from 0 to 1}
 \item{\code{asw}}{a numeric value presenting the average short-wave absorptivity of the occupant. Set to 0.7 by default}
\item{\code{posture}}{Default 'seated' list of available options 'standing', 'supine' or 'seated'}
 \item{\code{floorRef}}{a numeric value presenting floor refectance. It is assumed to be constant and equal to 0.6}
 \item{\code{erf}}{Solar gain to the human body using the Effective Radiant Field [W/m2]}
 \item{\code{delMrt}}{Delta mean radiant temperature [Degree C]}   
}
}
\usage{
data(solarGainValues)
}
\description{
Dataset with DIfferent Combination of Inputs to Calculate Solar Gain
}
\examples{
data(solarGainValues)
head(solarGainValues)
}
\keyword{datasets}
\keyword{internal}
