% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utciValue.R
\docType{data}
\name{utciValues}
\alias{utciValues}
\title{Dataset with DIfferent Combination of Inputs to Calculate UTCI}
\format{
A data frame with 81 rows and 5 variables:
\describe{
 \item{\code{ta}}{a numeric value presenting air temperature in [degree C]}
\item{\code{tr}}{a numeric value presenting mean radiant temperature in [degree C]}
 \item{\code{vel}}{a numeric value presenting air velocity in [m/s]}
\item{\code{rh}}{a numeric value presenting relative humidity [\%]}
\item{\code{utci_value}}{a numeric value presenting the UTCI value}
}
}
\usage{
data(utciValues)
}
\description{
Dataset with DIfferent Combination of Inputs to Calculate UTCI
}
\examples{
data(utciValues)
head(utciValues)
}
\keyword{datasets}
\keyword{internal}
