% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validate_UTCI}
\alias{validate_UTCI}
\title{Validation for Calculating UTCI}
\usage{
call calcUTCI(ta, tr, vel, rh), validate_UTCI will be called automatically
}
\arguments{
\item{ta}{a numeric value presenting air temperature in [degree C]}

\item{tr}{a numeric value presenting mean radiant temperature in [degree C]}

\item{vel}{a numeric value presenting air velocity in [m/s]}

\item{rh}{a numeric value presenting relative humidity [\%]}
}
\value{
Gives a warning message if any of the inputs are not numerical, null and out of rage. If the inputs are valid then nothing is done.
}
\description{
Functions to validate UTCI .
}
\details{
inputs have to be numerical and not null. Also, the values have to be of certain range.
}
\examples{
validate_UTCI(25, 25, 0.1, 50) #valid input
does nothing
validate_UTCI(FALSE, 50, 0.1, 50) #not numerical ta input
validate_UTCI(50, 50, 0.1, 50) #invalid input for ta, td
Warning Message: 50 is out of range. Has to be between 0 and 40"
                 50 is out of range. Has to be between 0 and 40
validate_UTCI(25, 25, 0.1, 150) #invalid input for rh
Warning Message: 150 is out of range. Has to be between 0 and 100"
}
\author{
Code implemented in to R by Shaomi Rahman
}
