% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAD.R
\name{calcAD}
\alias{calcAD}
\alias{ankledraft}
\alias{ad}
\title{Calculate PPD with Ankle Draft}
\usage{
calcAD(ta, tr, vel, rh, clo, met, v_ankle)

ad(ta, tr, vel, rh, clo, met, v_ankle)

ankledraft(ta, tr, vel, rh, clo, met, v_ankle)
}
\arguments{
\item{ta}{a numeric value presenting air temperature in [degree C]}

\item{tr}{a numeric value presenting mean radiant temperature in [degree C]}

\item{vel}{a numeric value presenting air velocity in [m/s]}

\item{rh}{a numeric value presenting relative humidity [\%]}

\item{clo}{a numeric value presenting clothing insulation level in [clo]}

\item{met}{a numeric value presenting metabolic rate in [met]}

\item{v_ankle}{air speed at the 0.1 m (4 in.) above the floor [m/s]}
}
\value{
Predicted Percentage of Dissatisfied occupants with ankle draft in [\%]

Acceptability in [boolean]
}
\description{
Function to calculate ankle draft using the function predicted percentage of dissatisfied.
}
\details{
Calculates the percentage of thermally dissatisfied people with the ankle draft (0.1 m) above floor level.This equation is only applicable for velocity < 0.2 m/s (40 fps)
}
\examples{
calcAD(25,25,0.2,50,0.5,1.2,0.3)
}
\author{
Shoaib Sarwar
}
