% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fctCE.R
\name{calcCE}
\alias{calcCE}
\alias{coolingeffect}
\alias{ce}
\title{Calculate Cooling Effect}
\usage{
calcCE(ta, tr, vel, rh, clo = 0.5, met = 1, wme = 0)

ce(ta, tr, vel, rh, clo = 0.5, met = 1, wme = 0)

coolingeffect(ta, tr, vel, rh, clo = 0.5, met = 1, wme = 0)
}
\arguments{
\item{ta}{a numeric value presenting air temperature in [degree C]}

\item{tr}{a numeric value presenting mean radiant temperature in [degree C]}

\item{vel}{a numeric value presenting air velocity in [m/s]}

\item{rh}{a numeric value presenting relative humidity [\%]}

\item{clo}{a numeric value presenting clothing insulation level in [clo]}

\item{met}{a numeric value presenting metabolic rate in [met]}

\item{wme}{a numeric value presenting external work in [met]}
}
\value{
ce - Cooling Effect in [degree C]
}
\description{
Function to calculate cooling effect using the function standard effective temperature.
}
\details{
The CE of the elevated air speed is the value that, when subtracted equally from both the average air temperature and the mean radiant temperature, yields the same SET under still air as in the first SET calculation under elevated air speed. The cooling effect is calculated only for air speed higher than 0.1 m/s.
}
\examples{
calcCE(25,25,0.3,50,0.5,1) # returns Cooling Effect: 1.3
}
\author{
Code implemented in to R by Shoaib Sarwar. Further contribution by Marcel Schweiker.
}
