% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPetSteady.R
\name{calcPetSteady}
\alias{calcPetSteady}
\title{Calculate the Steady Physiological Equivalent Temperature (PET)}
\usage{
calcPetSteady(
  tdb,
  tr,
  v,
  rh,
  met,
  clo,
  p_atm = 1013.25,
  position = 1,
  age = 23,
  sex = 1,
  weight = 75,
  height = 1.8,
  wme = 0
)
}
\arguments{
\item{tdb}{float, dry bulb air temperature, [degree]}

\item{tr}{float, mean radiant temperature, [degree]}

\item{v}{float, air speed, [m/s]}

\item{rh}{float, relative humidity, [%]}

\item{met}{float, metabolic rate, [met]}

\item{clo}{float, clothing insulation, [clo]}

\item{p_atm}{float, atmospheric pressure, default value 1013.25 [hPa]}

\item{position}{integer, position of the individual (1=sitting, 2=standing, 3=standing forced convection)}

\item{age}{integer, age in years, default 23}

\item{sex}{integer, gender (1 for male, 2 for female), default 1}

\item{weight}{float, body mass, [kg], default 75}

\item{height}{float, height, [m], default 1.8}

\item{wme}{float, external work, [W/(m2)], default 0}
}
\value{
PET, Steady-state PET under the given ambient conditions
}
\description{
The PET is calculated using the Munich Energy-balance Model for Individuals,
PET is defined as the air temperature at which, in a typical indoor setting,
the heat budget of the human body is balanced with the same core and skin
temperature as under the complex outdoor conditions to be assessed.
}
\details{
The PET was originally proposed by Hoppe. In 2018, Walther and Goestchel proposed a
correction of the original model, purging the errors in the PET calculation routine,
and implementing a state-of-the-art vapour diffusion model.
}
\examples{
pet_steady(tdb = 20, tr = 20, rh = 50, v = 0.15, met = 1.37, clo = 0.5)
}
\references{
Hoppe, P. (1999). The Physiological Equivalent Temperature - A Universal Index for the Biometeorological Assessment of the Thermal Environment.
Walther, G. and Goestchel, P. (2018). Improvement and Extension of the Munich Energy-balance Model for Individuals.
}
\author{
Code implemented into R by Chongyu Gan.
}
