% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPhs.R
\name{calcPhs}
\alias{calcPhs}
\title{Calculates the Predicted Heat Strain (PHS) based on ISO 7933:2004}
\usage{
calcPhs(
  tdb,
  tr,
  v,
  rh,
  met,
  clo,
  posture,
  wme = 0,
  limit_inputs = TRUE,
  i_mst = 0.38,
  a_p = 0.54,
  drink = 1,
  weight = 75,
  height = 1.8,
  walk_sp = 0,
  theta = 0,
  acclimatized = 100,
  duration = 480,
  f_r = 0.97,
  t_sk = 34.1,
  t_cr = 36.8,
  t_re = FALSE,
  t_cr_eq = FALSE,
  t_sk_t_cr_wg = 0.3,
  sweat_rate = 0
)
}
\arguments{
\item{tdb}{float or numeric vector. Dry bulb air temperature, default in Celsius}

\item{tr}{float or numeric vector. Mean radiant temperature, default in Celsius}

\item{v}{float or numeric vector. Air speed, default in [m/s]}

\item{rh}{float or numeric vector. Relative humidity, [\%]}

\item{met}{float or numeric vector. Metabolic rate, [W/m2]}

\item{clo}{float or numeric vector. Clothing insulation, [clo]}

\item{posture}{integer. Numeric value representing the posture of the person
(1 = sitting, 2 = standing, 3 = crouching)}

\item{wme}{float or numeric vector. External work, [W/m2], default 0}

\item{limit_inputs}{logical, default TRUE. If TRUE, returns NA when inputs are
outside the standard applicability limits (15 < tdb < 50 Celsius, 0 < tr < 60 Celsius,
0 < vr < 3 [m/s], 100 < met < 450 [met], 0.1 < clo < 1).}

\item{i_mst}{float, default 0.38. Static moisture permeability index, [dimensionless]}

\item{a_p}{float, default 0.54. Fraction of body surface covered by reflective clothing, [dimensionless]}

\item{drink}{integer, default 1. Whether workers can drink freely (1 = yes, 0 = no)}

\item{weight}{float, default 75. Body weight, [kg]}

\item{height}{float, default 1.8. Height, [m]}

\item{walk_sp}{float, default 0. Walking speed, [m/s]}

\item{theta}{float, default 0. Angle between walking direction and wind direction, [Celsius]}

\item{acclimatized}{integer, default 100. 100 if acclimatized subject, 0 otherwise}

\item{duration}{integer, default 480. Duration of the work sequence, [minutes]}

\item{f_r}{float, default 0.97. Emissivity of reflective clothing, [dimensionless]}

\item{t_sk}{float, default 34.1. Mean skin temperature when worker starts working, Celsius}

\item{t_cr}{float, default 36.8. Mean core temperature when worker starts working, Celsius}

\item{t_re}{float, default FALSE. Mean rectal temperature when worker starts working, Celsius}

\item{t_cr_eq}{float, default FALSE. Mean core temperature as a function of the metabolic rate when worker starts working,}

\item{t_sk_t_cr_wg}{fraction of the body mass at the skin temperature}

\item{sweat_rate}{float, default 0. Initial sweat rate}
}
\value{
\describe{
  \item{t_re}{Rectal temperature in Celsius}
  \item{t_sk}{Skin temperature in Celsius}
  \item{t_cr}{Core temperature in Celsius}
  \item{t_cr_eq}{Core temperature as a function of metabolic rate, in Celsius}
  \item{t_sk_t_cr_wg}{Fraction of body mass at the skin temperature}
  \item{d_lim_loss_50}{Maximum allowable exposure time for water loss for a mean subject, in minutes}
  \item{d_lim_loss_95}{Maximum allowable exposure time for water loss for 95\% of the working population, in minutes}
  \item{d_lim_t_re}{Maximum allowable exposure time for heat storage, in minutes}
  \item{water_loss_watt}{Maximum water loss in watts}
  \item{water_loss}{Maximum water loss in grams}
}
}
\description{
Calculates the Predicted Heat Strain (PHS) based on ISO 7933:2004
}
\details{
The PHS model predicts:
- Heat by respiratory convection and evaporation
- Steady-state mean skin temperature and instantaneous skin temperature
- Heat accumulation associated with metabolic rate
- Maximum evaporative heat flow at the skin surface
- Predicted sweat rate and evaporative heat flow
- Rectal temperature
}
\examples{
# Example usage of phs function
calcPhs(tdb = 40, tr = 40, rh = 33.85, v = 0.3, met = 150, clo = 0.5, posture = 2, wme = 0)
}
\author{
Code implemented into R by Chongyu Gan.
}
