% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfField.R
\docType{data}
\name{dfField}
\alias{dfField}
\title{Field data example}
\format{
A data frame with 156 rows and 9 variables:
\describe{
  \item{ta}{air temperature [degree C]}
  \item{tr}{radiant temperature [degree C] - same as ta}
  \item{rh}{relative humidity [\%]}
  \item{trm}{running mean outdoor temperature [degree C]}
  \item{clo}{clothing insulation level [CLO]}
  \item{tao}{outdoor air temperature [degree C]}
  \item{vel}{indoor air velocity [m/s]}
  \item{met}{metabolic rate [MET]}
  \item{asv}{actual thermal sensation vote on ASHRAE scale [ ]}
}
}
\usage{
data(dfField)
}
\description{
Randomly sampled data from a field study campaign
with data from 156 samples. For further description, 
see the reference given.
}
\examples{
data(dfField)
head(dfField)
}
\references{
Schweiker, M. and Wagner, A. Exploring potentials and limitations of the adaptive thermal heat balance framework. Proceedings of 9th Windsor Conference: Making Comfort Relevant Cumberland Lodge, Windsor, UK, 2016.
(\href{https://www.windsorconference.com}{nceub})
}
\keyword{datasets}
