% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfUtcivalues.R
\docType{data}
\name{dfUTCIValues}
\alias{dfUTCIValues}
\title{Dataset with Different Combinations of Inputs to Calculate UTCI}
\format{
A data frame with 81 rows and 5 variables:
\describe{
 \item{\code{ta}}{a numeric value presenting air temperature in [degree C]}
\item{\code{tr}}{a numeric value presenting mean radiant temperature in [degree C]}
 \item{\code{vel}}{a numeric value presenting air velocity in [m/s]}
\item{\code{rh}}{a numeric value presenting relative humidity [\%]}
\item{\code{utci}}{a numeric value presenting the UTCI value}
}
}
\usage{
data(dfUTCIValues)
}
\description{
Dataset with Different Combinations of Inputs to Calculate UTCI
}
\examples{
data(dfUTCIValues)
head(dfUTCIValues)
}
\seealso{
see also \code{\link{calcUTCI}}
}
\keyword{datasets}
